chore: moved csvColumns & examples to internal
